requirements.txt
# All pins reflect fixes for CVEs found by `pip-audit` (Apr 2026 audit).
fastapi==0.136.1                 # latest, pulls starlette >= 0.49 (CVE-2025-62727 patched)
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.26         # CVE-2024-53981 / CVE-2026-24486 / CVE-2026-40347
PyJWT==2.12.0                    # CVE-2026-32597
python-dotenv==1.2.2             # CVE-2026-28684
slowapi==0.1.9
bcrypt==4.2.1
redis==5.2.1
httpx==0.28.1
pyyaml==6.0.2
scikit-learn==1.5.2              # PYSEC-2024-110
numpy==2.2.0
scipy==1.14.1
taxii2-client==2.3.0
stix2==3.0.1
prometheus_client==0.21.1
# Database: ORM + migrations (PostgreSQL production / SQLite demo)
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10          # PostgreSQL driver (optional — only needed with PostgreSQL)
# OIDC/SSO: JWKS validation + JWT decoding for enterprise SSO
authlib>=1.3
# Field-level encryption (AES-256-GCM, HKDF key derivation)
cryptography>=44.0
# Background jobs: Arq (Redis-backed, asyncio-native, ~1k LoC, 0 CVE)
arq==0.26.3
# OpenTelemetry (optional — enable with OTEL_ENABLED=true)
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
